requirements.txt
streamlit>=1.32.0
easyocr>=1.7.0
plotly>=5.18.0
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
openpyxl>=3.1.0
